Refactor(PurchasePrice): PurchasePrice 클래스 구조 변경
- price받는 생성자 추가
- validator 추출
