+++ src/main/java/lotto/validator/PurchasePriceValidator.java
@@ -1,0 +1,20 @@
+package lotto.validator;
+
+public class PurchasePriceValidator {
+
+    private boolean isOverMinPrice() {
+        return true;
+    }
+
+    private void checkOverMinPrice() {
+
+    }
+
+    private boolean isRemainderZero() {
+        return true;
+    }
+
+    private void checkRemainderZero() {
+
+    }
+}
